perf(table): 优化双表格-默认使用单表格
